# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quantum FX Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/earnings-chart:
    get:
      operationId: getEarningsChart
      tags: [dashboard]
      parameters:
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Earnings chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartPoint"
...
  /cashier/deposit:
    post:
      operationId: createDeposit
      tags: [cashier]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cashier/deposit/session:
    post:
      operationId: createDepositSession
      tags: [cashier]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepositSessionInput"
      responses:
        "201":
          description: Deposit session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositSession"
  /cashier/deposit/session/{id}:
    get:
      operationId: getDepositSession
      tags: [cashier]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositSession"
  /cashier/deposit/session/{id}/txid:
    post:
      operationId: submitDepositTxid
      tags: [cashier]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTxidInput"
      responses:
        "200":
          description: TXID submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositSession"
...
  /cashier/withdraw:
    post:
      operationId: createWithdrawal
      tags: [cashier]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "201":
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cashier/transactions:
    get:
      operationId: listTransactions
      tags: [cashier]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, deposit, withdrawal]
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ChartPoint:
      type: object
      properties:
        date:
          type: string
        label:
          type: string
        profit:
          type: number
        cumulative:
          type: number
      required: [date, label, profit, cumulative]
...
    DepositInput:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        walletAddress:
          type: ["string", "null"]
      required: [amount, paymentMethod]
    WithdrawInput:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        walletAddress:
          type: string
      required: [amount, paymentMethod, walletAddress]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        createdAt:
          type: string
        walletAddress:
          type: ["string", "null"]
      required: [id, type, amount, status, paymentMethod, createdAt]
...
    DepositSession:
      type: object
      properties:
        id: { type: integer }
        status: { type: string }
        amount: { type: number }
        paymentMethodId: { type: string }
        paymentMethodName: { type: string }
        network: { type: string }
        depositAddress: { type: string }
        txid: { type: ["string", "null"] }
        confirmations: { type: integer }
        requiredConfirmations: { type: integer }
        expiresAt: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, status, amount, paymentMethodId, paymentMethodName, network, depositAddress, confirmations, requiredConfirmations, expiresAt, createdAt, updatedAt]
    CreateDepositSessionInput:
      type: object
      properties:
        amount: { type: number }
        paymentMethodId: { type: string }
      required: [amount, paymentMethodId]
    SubmitTxidInput:
      type: object
      properties:
        txid: { type: string }
      required: [txid]
...